Update coloring logic for unique portals
Normally this plugin adds a red highlight if you haven't captured a portal, and yellow if you've only visited it. I don't find visited a useful stat to track so I'm ignoring that. Now it'll show portals in a conspicuous magenta color if you can go capture it right now to add it to your Pioneer badge progress. It shows portals in a less visible red if you've never captured it and it's currently owned by the same faction you're on, so you can't capture it yourself. Also I'm making neutral portals white because I think that looks better.